Refactoring: Utilize the shared ProgressInfo also in Music app proper
There's no reason to duplicate the progress component logic using the
AngularJS syntax. We now have a custom element directive <progress-info>
which is just a thin wrapper for the ProgressInfo class used also in the
embedded Files player and the Dashboard widget. The shared ProgressInfo
communicates directly with the PlayerWrapper and needs no other boiler-plate
in the application code.