Initial support for RTL language layout
In NC 31 and later, the UI is mirrored when a right-to-left language like
Hebrew or Arabic is selected as the UI language. The Music app UI was
previously totally broken and unusable in this case. With this initial set of
changes, the UI becomes usable but a few problems still remain:

- As convention, the order of the playback buttons should not be inverted
  with the rest of the UI (see e.g.
https://m2.material.io/design/usability/bidirectionality.html#mirroring-elements)

- The UI doesn't currently react to the app-view size changes on the RTL
  layout. This is apparently a bug in the jquery-resize plugin which we use.