Fix media control seek forward/backward not working on Edge
It seems that recent versions of Chrome are not actually using the action
handlers "seekforward" and "seekbackward" which had the bug. Instead, it has
to be implementing the seeking more directly on the HTML audio element.
Chrome also provides a possibility to seek to any point within the file which
is not supported in the standard MediaSession API.

On the other hand, the latest Edge version 142 now provides similar media
controls as Chrome had in the past. This actually uses those "seekforward"
and "seekbackward" handlers and this revealed the bug.