Return rudimentary support for IE after fd49b09
Internet Explorer doesn't support the properties related to bi-directional UI
like "inset-inline-start" or "margin-inline-end". Hence, these need to be
overridden on IE with the more traditional alternatives "left" and
"margin-right".

We had already previously lot of layout issues on IE and the point of this
commit was not to make the UI work perfectly on IE. But at least it's now
again somewhat usable there.