Overlay time text on progress bar in the controls bar two-line layout
Previously, the space reserved for the progress text was too narrow when the
length of the recording was more than one hour. Also, the progress bar was
very narrow on portrait phone screen. As an added benefit, we can now keep
the progress bar visible down to the narrowest possible windows.