Use the same overlayed progress text color in Dashboard as on Music app proper
The previous color selection of black font on the "primary color" progress
bar was difficult to read especially on the light mode, depending a bit on
the color theme used. Using a light font, instead, shows perfectly on the
primary color but the challenge is to make it readable on the light gray
"empty" area of the progress bar. However, with a black shadow, it's somewhat
readable also there and overall the compromise is better.