Make the progress bar background a tiny bit darker on dark mode
It's not possible to (easily) define CSS rules targeting the dark mode. To
achieve our goal, the background was made a little bit translucent. Hence, it
appears a bit darker on dark background but almost identical to the old color
on the bright background.

This will still improve the readability of the progress text overlaid on the
background of the progress bar. It used to be a bit more difficult to read on
the dark mode compared to the bright mode but now there's no longer
significant difference. The specs of the display device may matter, though.