Keep playback buttons in LTR order even on RTL page layout
The common convention is to not mirror the playback buttons even when the UI
language is a one using the right-to-left layout; see e.g.
https://m2.material.io/design/usability/bidirectionality.html#mirroring-elements.

This commit touches only the Music app proper, the Files player and the
Dashboard widget should still be fixed the same way.

Previously, both padding of #controls and margin of #play-controls was
affecting the placement of the playback control buttons. This was simplified
so that only the padding in #controls dictates this. The padding doesn't
affect the rest of the contents of the #controls div since everything else
has absolute positions.