Tweak the alignment of the play button
To look balanced, an icon like the play icon must not be technically
center-aligned. Instead, the "center of mass" of the icon should be in the
middle point (between skip-next and skip-prev buttons).

To achieve this, some margin space was added to the play-big.svg and the
graphic of the icon was right-aligned in the canvas. Also pause-big.svg and
stop.svg were modified to the same width (as they occupy the same slot on the
UI) but their graphics are center-aligned in the canvas. Then, CSS rules were
modified to cope with the new shape of these icons while keeping all the
other elements except the play icon visually in their previous places and
sizes.

The modification was made on all the three players: Music app proper,
embedded Files player, and Dashboard widget.