Fix loading icon being misplaced on Dashboard widget on RTL
The margin given by the CSS of the NC 32.0.1 server for the loading icon is
`margin: -16px 0 0 -16px;`.

That is, margin-top and margin-left are set to a negative values. This works
for the LTR layout bu not for RTL. However, all the margins to the same value
works fine both on LTR and RTL.