Use marginless play/pause icon on list items and headers
The margin added in 4d73926 to play-big.svg and pause-big.svg caused these
icons on the list items and headers to become distorted with wrong aspect
ratio. Provide another pair of icons without any margins to be used in these
places. The new icons also have a smaller nominal size, somewhat reflecting
the size in which the icons are used.