Identify playlist radio streams from the Content-Type header instead of URL
Previously, the radio streaming URL was handled as a playlist if the path
part of the URL ended with .pls, .m3u, or .m3u8 (case-insensitive). This
failed on some radio stations since nothing guarantees that the URL actually
contains the file name. Now we always fetch headers from the stream URL and
check the MIME contained in the Content-Type header.