Subsonic: Provide a non-dummy implementation for `getNowPlaying`
Replace the empty/dummy implementation added in b23a56a with a one returning
some valid data. As usual, we don't provide other users' data even though
that's how the endpoint is specified because we don't want to leak any user
data without an explicit consent. The song shown in the response is based on
the scrobbling messages from the clients.