Downgrade phpunit back to v8 with help of composer-bin-plugin
The phpunit v9 could not run the integration tests on PHP 7.4 and had to
be downgraded back to v8. To enable this while keeping coding-standard
up-to-date, whe had to solve the "dependency hell". The solution suggested
in the Nextcloud development guide was applied and a parallel vendor
environment was installed for phpunit with the help of the composer plugin
bamarni/composer-bin-plugin.