Tweak how artist name is separated from track title in "Now playing"
In case there are multiple instances of the separator string " - " in the
title broadcasted by the radio channel, the matching logic now assumes that
the first one separates the artist from the track title and the next one(s)
belong to the track title. This is achieved with the "lazy" repeat quantifier
"+?".

While the regex previously operated entirely in the greedy mode, it split the
artist from the track title using the last found instance of " - ". This was
seen to fail with the title "Parker McCollum - What Kinda Man - Radio Edit".
Indeed there certainly exists a counter example where splitting from the last
separator would be the right thing to do. However, the assumption is that
those cases would be less common.