Handle race conditions on radio Last.fm data updating
Previously, it could happen that the Last.fm data fetching got triggered
multiple times back-to-back with different track/artist names. In this case,
it was random, which one of the fetches finished last and was left visible on
the UI. This could happen especially when switching back and forth between
stations and the current song of one of the stations changed.

With the changed logic, we may still sometimes fetch info for many songs
back-to-back but only the data of the actual current track should get
rendered. Also, we now store the fetched Last.fm data per-station,
potentially reducing queries to the cloud server and eventually to the Last.fm
server.