In Settings view, don't show the "Admin has configured scrobbling" text
Never show the text "The server administrator has configured your server for
scrobbling." This was redundant since it's clear that the scrobbling is
configured if the user is able to connect or disconnect the scrobbler. The
opposite text "The server administrator can configure your server for
scrobbling, see the <strong>Admin</strong> section below." is valuable and
needed, though.

Also, tweaked the layout a tiny bit by moving the scrobbler
Connect/Disconnect buttons off the div containing the instruction texts. With
this, we get a small margin between the text and the buttons which looks much
nicer.