Fix Advanced Search common rules layout on RTL
The order of the common parameter elements
"Search for <type selection> <match mode> ordering results <order selection>
limiting to <count selection>" was bizarre, not being fully RTL or LTR but
somehow mixed so that all labels and matching selections were not next to
each other.

Now each label-selection pair is placed in its own <div>, ensuring that they
stay together. Also, the labels are now <label> elements instead of <span>
elements and have the `for` attribute to help screen readers.