In GitHub front page, show CI status from current repository
Previously, the status was always shown for the fork paulijar/music,
regardless of which fork was being viewed. Note that the relative links need
to start with `../../` because GitHub resolves them relative to the path
`github.com/<owner>/<repo>/blob/<branch>`

The Scrutinizer status is still shown for the original fork owncloud/music
since there we can't just use the relative link to achieve the dynamic
behavior.