Subsonic: Fix `id` formatting on property `artists` of album results
There are already several OpenSubsonic-compatible clients utilizing the
`artists` property added in Music v2.5.0. On many such clients, the links
from album view to the contained artists were broken because of this bug.
This was seen at least on Supersonic, Feishin, and Airdrome but probably
there are others as well.